pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-sandbox"
version = "0.1.0"
description = "Sandbox starter project for agentic AI experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "openai>=2.0.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "typer>=0.12.3",
]

[project.scripts]
agentic-sandbox = "agentic_sandbox.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
